requirements.txt
requests
asyncio
pydantic
python-dotenv
scipy
sounddevice
shazamio
python3-discogs-client
musicbrainzngs